+++ unstructured_inference/constants.py
@@ -13,9 +13,17 @@ class Source(Enum):
     CHIPPER = "chipper"
     CHIPPERV1 = "chipperv1"
     CHIPPERV2 = "chipperv2"
-    PDFMINER = "pdfminer"
+    CHIPPERV3 = "chipperv3"
     MERGED = "merged"
     SUPER_GRADIENTS = "super-gradients"
+
+
+CHIPPER_VERSIONS = (
+    Source.CHIPPER,
+    Source.CHIPPERV1,
+    Source.CHIPPERV2,
+    Source.CHIPPERV3,
+)
 
 
 class ElementType: